Modify the commandline so that it aborts after completion.
A flag has been added in Commander.cs that is set when the writing of the results is done.
The method startEvolution() only returns after that flag is set.
This allows the removal of the Extra Console.ReadLine() in the file Program.cs and therefore the commandline exits without additional input.