Fix error while generating Interactions
If there where no more interactions to create for the current order, a interaction int the next higher order was allways create.
Even if there where no interactions in the next higher order.
This is simply fixed by increasing the order for as long as there are no interactions in the current order.